Split `main.py` and `api.py`
